Okay, more rust junk

+++ src/rust/src/continuity.rs
@@ -2,5 +2,5 @@
 pub struct Continuity {
     pub pos:   (i8, i8),
     pub neg:   (i8, i8),
-    pub score: u32, // 24-bit UNORM, raw
+    pub score: f32, // 24-bit UNORM, raw
 }
